Move BattleResults to games.strategy.triplea.delegate
