+++ game-core/src/main/java/games/strategy/triplea/delegate/BattleDelegate.java
@@ -44,7 +44,6 @@ import games.strategy.triplea.delegate.dataObjects.BattleListing;
 import games.strategy.triplea.delegate.dataObjects.BattleRecord;
 import games.strategy.triplea.delegate.remote.IBattleDelegate;
 import games.strategy.triplea.formatter.MyFormatter;
-import games.strategy.triplea.oddsCalculator.ta.BattleResults;
 import games.strategy.triplea.player.ITripleAPlayer;
 import games.strategy.util.CollectionUtils;
 import games.strategy.util.IntegerMap;
