+++ game-core/src/main/java/games/strategy/triplea/delegate/dataObjects/BattleRecord.java
@@ -7,7 +7,7 @@ import games.strategy.engine.data.PlayerID;
 import games.strategy.engine.data.SerializationProxySupport;
 import games.strategy.engine.data.Territory;
 import games.strategy.triplea.delegate.IBattle.BattleType;
-import games.strategy.triplea.oddsCalculator.ta.BattleResults;
+import games.strategy.triplea.delegate.BattleResults;
 
 /**
  * The Purpose of this class is to record various information about combat,
