+++ game-core/src/main/java/games/strategy/triplea/odds/calculator/OddsCalculator.java
@@ -16,7 +16,7 @@ import games.strategy.engine.framework.GameDataUtils;
 import games.strategy.triplea.delegate.BattleTracker;
 import games.strategy.triplea.delegate.GameDelegateBridge;
 import games.strategy.triplea.delegate.MustFightBattle;
-import games.strategy.triplea.oddsCalculator.ta.BattleResults;
+import games.strategy.triplea.delegate.BattleResults;
 
 class OddsCalculator implements IOddsCalculator, Callable<AggregateResults> {
   private GameData gameData;
